agregado productsService
